chore: debug and UI fixes for positions and orders

+++ frontend/src/components/FullOrderEntry.tsx
@@ -302,7 +302,7 @@ export default function FullOrderEntry() {
               type="number"
               min="1"
               value={orderData.quantity}
-              onChange={(e) => setOrderData({...orderData, quantity: parseInt(e.target.value) || 1})}
+              onChange={(e) => setOrderData({...orderData, quantity: parseFloat(e.target.value) || 1})}
             />
           </div>
           <div>
